requirements.txt
anthropic>=0.40.0
supabase>=2.0.0
feedparser>=6.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
Pillow>=10.0.0
google-api-python-client>=2.100.0
google-auth>=2.25.0
python-dotenv>=1.0.0
structlog>=23.0.0
pyyaml>=6.0.0
schedule>=1.2.0
pytz>=2024.1
pytest>=8.0.0
